+++ R/valueButton.R
@@ -1,6 +1,6 @@
 #' Take value from selected UI element
 #'
-#' The components creates button that allows to take any value (or attribute) sourced
+#' The components creates button or link that allows to take any value (or attribute) sourced
 #' from DOM element existing in Shiny application and pass it to application server.
 #'
 #' @examples
@@ -25,18 +25,15 @@
 #'   For nested properties use `.`, e.g. `style.width` or `navigator.appName`.
 #' @param icon Icon included in button.
 #' @param width Width of the button.
-#' @param ... Extra attributes passed to button.
+#' @param ... Extra attributes passed to `button` or `a` tag.
 #'
 #' @return A `shiny.tag` class object defining html structure of the button.
+#' @name valueButton
 #' @export
 valueButton <- function(inputId, label, selector, attribute = "value", icon = NULL, width = NULL, ...) {
 
   shiny::tagList(
-    shiny::singleton(
-      shiny::tags$head(
-        shiny::tags$script(type = "text/javascript", src = "shinyGizmo/valuebutton.js")
-      )
-    ),
+    value_button_dependency,
     shiny::tags$button(
       id = inputId,
       style = htmltools::css(width = shiny::validateCssUnit(width)),
@@ -51,4 +48,28 @@ valueButton <- function(inputId, label, selector, attribute = "value", icon = NU
   )
 }
 
+#' @rdname valueButton
+#' @export
+valueLink <- function(inputId, label, selector, attribute = "value", icon = NULL, ...) {
+
+  shiny::tagList(
+    value_button_dependency,
+    shiny::tags$a(
+      id = inputId,
+      class = "value-button",
+      href = "#",
+      `data-val` = NULL,
+      `data-value-attribute` = attribute,
+      `data-selector` = selector,
+      list(icon, label),
+      ...
+    )
+  )
+}
+
+value_button_dependency <- shiny::singleton(
+  shiny::tags$head(
+    shiny::tags$script(type = "text/javascript", src = "shinyGizmo/valuebutton.js")
+  )
+)
 # todo implement update method