+++ NEWS.md
@@ -1,5 +1,6 @@
-# shinyGizmo 0.1.1
+# shinyGizmo (development version)
 
+* Add `valueLink` component.
 * Make `showModal` and `hideModal` work with modules.
 
 # shinyGizmo 0.1